separated network queryer logic from queryer utils
